headless merge前の責務境界を整理
