+++ internal/agent/headless_runner.go
@@ -2,7 +2,6 @@ package agent
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"io"
 	"strings"
@@ -11,9 +10,6 @@ import (
 	"github.com/susugadx/xelyon-cli/internal/api"
 	"github.com/susugadx/xelyon-cli/internal/config"
 	"github.com/susugadx/xelyon-cli/internal/prompt"
-	"github.com/susugadx/xelyon-cli/internal/toolruntime"
-	"github.com/susugadx/xelyon-cli/internal/tools"
-	"github.com/susugadx/xelyon-cli/internal/tools/subagent"
 	"github.com/susugadx/xelyon-cli/internal/uiruntime"
 )
 
@@ -144,332 +140,4 @@ func (r *headlessRunner) run(ctx context.Context) *HeadlessResult {
 		return r.loopLimitResult(maxIterations)
 	}
 	return r.successResult()
-}
-
-func (r *headlessRunner) requestAssistantResponse(ctx context.Context, iteration int) (string, error) {
-	timeout := time.Duration(r.agent.cfg().APIRetry.Timeout) * time.Second
-	if timeout <= 0 {
-		timeout = 3600 * time.Second
-	}
-	reqCtx, cancel := context.WithTimeout(ctx, timeout)
-	defer cancel()
-
-	// ツールループ初回のみ Project Map を更新。ループ中の再生成はキャッシュを破壊する。
-	if iteration == 0 {
-		r.agent.refreshProjectPromptIfDirty(r.query)
-	}
-
-	effectivePrompt := r.agent.SystemPrompt
-	var runtimeDirectives []string
-	if strings.TrimSpace(r.agent.cfg().SubAgentPrompt) == "" {
-		runtimeDirectives = r.agent.pendingRuntimeDirectives()
-		effectivePrompt = r.agent.normalModeSystemPromptForRequestWithDirectives(reqCtx, r.query, iteration == 0, runtimeDirectives)
-	}
-	requestCtx := r.agent.prepareResponseContextForPrompt(r.agent.requestContext(reqCtx), effectivePrompt)
-	if iteration == 0 && r.options.Image != nil {
-		requestCtx, history := r.agent.providerFacingHistoryExcludingLatestMessageForRequest(requestCtx)
-		response, err := r.provider.ChatWithImage(requestCtx, effectivePrompt, history, r.query, r.options.Image, r.model)
-		if err != nil {
-			return "", err
-		}
-		r.agent.recordResponseContextForPrompt(effectivePrompt)
-		r.agent.markRuntimeDirectivesDelivered(runtimeDirectives)
-		return response, nil
-	}
-
-	requestCtx, history := r.agent.providerFacingHistoryForRequest(requestCtx)
-	response, err := r.provider.ChatWithTools(requestCtx, effectivePrompt, history, r.model)
-	if err != nil {
-		return "", err
-	}
-	r.agent.recordResponseContextForPrompt(effectivePrompt)
-	r.agent.markRuntimeDirectivesDelivered(runtimeDirectives)
-	return response, nil
-}
-
-func (r *headlessRunner) handleAssistantResponse(ctx context.Context, response string) bool {
-	parsedCalls := r.agent.parseToolCalls(response)
-	if r.options.ReadOnly {
-		parsedCalls = append(parsedCalls, headlessReadOnlyMCPXMLToolAttempts(response)...)
-	}
-	assignHeadlessRescueToolCallIDs(parsedCalls)
-
-	// ツール呼び出しがなければ最終レスポンスとして終了
-	if len(parsedCalls) == 0 {
-		r.finalReply = response
-		r.runFinalChecksIfNeeded(ctx)
-		return true
-	}
-
-	appendHeadlessToolCallsToHistory(r.agent, response, parsedCalls)
-	r.executeToolCalls(ctx, parsedCalls)
-
-	// 最大イテレーション到達時のフォールバック
-	r.finalReply = response
-	return false
-}
-
-func (r *headlessRunner) executeToolCalls(ctx context.Context, calls []*tools.ToolCall) {
-	for _, tc := range calls {
-		r.executeToolCall(ctx, tc)
-	}
-}
-
-func (r *headlessRunner) executeToolCall(ctx context.Context, tc *tools.ToolCall) {
-	toolCount := len(r.toolCalls) + 1
-	subagent.EmitEvent(ctx, subagent.SubAgentEvent{
-		Tool:      tc.Tool,
-		Phase:     "start",
-		FilePath:  extractToolFilePath(tc),
-		ToolIndex: toolCount,
-	})
-
-	execResult, denied := r.readOnlyDeniedToolResult(tc)
-	if denied {
-		r.readOnlyViolation = true
-	} else {
-		execResult = r.agent.executeQuietToolResult(ctx, tc, strings.NewReader(""), io.Discard, io.Discard, true)
-		r.agent.recordSkillActivationFromToolResult(tc, execResult.Result, execResult.Error)
-		r.agent.mutationTracker().recordExecutedToolResult(tc, execResult, true)
-	}
-	output := execResult.Result
-
-	success := isHeadlessToolCallSuccessForTool(tc.Tool, execResult)
-	if r.agent.Stats != nil {
-		r.agent.Stats.AddToolExecution(tc.Tool)
-	}
-	if summary, ok := newHeadlessCommandSummary(tc, execResult); ok {
-		r.commands = append(r.commands, summary)
-	}
-
-	r.toolCalls = append(r.toolCalls, ToolCallResult{
-		Tool:    tc.Tool,
-		Args:    tc.Args,
-		Output:  output,
-		Success: success,
-	})
-
-	event := subagent.SubAgentEvent{
-		Tool:      tc.Tool,
-		Phase:     "end",
-		FilePath:  extractToolFilePath(tc),
-		Success:   success,
-		Output:    truncateEventOutput(output, 200),
-		ToolIndex: toolCount,
-	}
-	if tc.Tool == "str_replace" {
-		event.OldStr = tc.Args["old_str"]
-		event.NewStr = tc.Args["new_str"]
-	}
-	subagent.EmitEvent(ctx, event)
-	appendHeadlessToolResultToHistory(r.agent, tc, output)
-}
-
-// isHeadlessToolCallSuccess は headless 実行におけるツール結果の成功判定を返す。
-// tools 層の共通契約（Error prefix 判定と Error flag）に揃えることで、
-// 先頭空白付きの "Error:" を失敗として扱いつつ、文中の "Error:" は許容する。
-func isHeadlessToolCallSuccess(execResult tools.ExecutionResult) bool {
-	return !execResult.Error && !tools.IsErrorResult(execResult.Result)
-}
-
-func isHeadlessToolCallSuccessForTool(toolName string, execResult tools.ExecutionResult) bool {
-	if !isHeadlessToolCallSuccess(execResult) {
-		return false
-	}
-	if toolName == subagent.WaitAgentToolName && headlessWaitAgentResponseHasFailure(execResult.Result) {
-		return false
-	}
-	return true
-}
-
-func headlessWaitAgentResponseHasFailure(output string) bool {
-	var response subagent.WaitResponse
-	if err := json.Unmarshal([]byte(output), &response); err != nil {
-		return false
-	}
-	if isHeadlessWaitAgentFailureStatus(response.Status) {
-		return true
-	}
-	for _, result := range response.Results {
-		if isHeadlessWaitAgentFailureStatus(result.Status) {
-			return true
-		}
-	}
-	return false
-}
-
-func isHeadlessWaitAgentFailureStatus(status string) bool {
-	switch strings.ToLower(strings.TrimSpace(status)) {
-	case "", "completed":
-		return false
-	default:
-		return true
-	}
-}
-
-func (r *headlessRunner) successResult() *HeadlessResult {
-	duration := time.Since(r.startedAt).Milliseconds()
-	result := r.attachSummary(attachHeadlessStats(r.agent, NewSuccessResult(r.provider.Name(), r.model, r.finalReply, r.toolCalls, duration)))
-	if r.cancelledErr != nil {
-		return promoteHeadlessCancelledResult(result, r.cancelledErr)
-	}
-	if r.finalCheckFailed {
-		return promoteHeadlessFinalCheckFailedResult(result)
-	}
-	if r.options.FailOnToolError && r.readOnlyViolation {
-		return promoteHeadlessReadOnlyViolationResult(result)
-	}
-	if r.options.FailOnToolError && hasFailedHeadlessToolCall(r.toolCalls) {
-		return promoteHeadlessToolErrorResult(result)
-	}
-	return result
-}
-
-func (r *headlessRunner) errorResult(errType, errMsg string) *HeadlessResult {
-	duration := time.Since(r.startedAt).Milliseconds()
-	return r.attachSummary(attachHeadlessStats(r.agent, NewErrorResult(r.provider.Name(), r.model, errType, errMsg, duration)))
-}
-
-func (r *headlessRunner) loopLimitResult(limit int) *HeadlessResult {
-	duration := time.Since(r.startedAt).Milliseconds()
-	result := NewToolLoopLimitResult(r.provider.Name(), r.model, limit, r.toolCalls, duration)
-	if r.options.FailOnToolError && r.readOnlyViolation {
-		result = promoteHeadlessReadOnlyViolationResult(result)
-	}
-	return r.attachSummary(attachHeadlessStats(r.agent, result))
-}
-
-func attachHeadlessStats(agent *Agent, result *HeadlessResult) *HeadlessResult {
-	if agent == nil || result == nil || agent.Stats == nil {
-		return result
-	}
-
-	agent.statsMu.Lock()
-	defer agent.statsMu.Unlock()
-
-	result.Tokens = &TokenUsage{
-		Input:    agent.Stats.InputTokens,
-		Cached:   agent.Stats.CachedInputTokens,
-		Output:   agent.Stats.OutputTokens,
-		Thinking: agent.Stats.ThinkingTokens,
-		Total:    agent.Stats.TotalTokens(),
-	}
-	if agent.Stats.WebSearchCalls > 0 {
-		result.WebSearch = &WebSearchUsage{
-			Calls:        agent.Stats.WebSearchCalls,
-			FeeEstimate:  agent.Stats.WebSearchCost,
-			ResultTokens: agent.Stats.WebSearchResultTokens,
-		}
-	}
-	estimate := agent.Stats.EstimatedCostEstimateForConfig(agent.cfg())
-	result.Cost = estimate.Cost
-	result.PricingUnavailable = estimate.PricingUnavailable
-	return result
-}
-
-func hasFailedHeadlessToolCall(toolCalls []ToolCallResult) bool {
-	for _, call := range toolCalls {
-		if !call.Success {
-			return true
-		}
-	}
-	return false
-}
-
-func assignHeadlessRescueToolCallIDs(toolCalls []*tools.ToolCall) {
-	for i, tc := range toolCalls {
-		if tc == nil || tc.ID != "" {
-			continue
-		}
-		toolCalls[i].ID = fmt.Sprintf("call_rescue_%03d", i+1)
-	}
-}
-
-func appendHeadlessToolCallsToHistory(agent *Agent, response string, toolCalls []*tools.ToolCall) {
-	if agent == nil || len(toolCalls) == 0 {
-		return
-	}
-
-	explanation, _ := extractExplanationAndTool(response)
-	reasoningContent := agent.getLastReasoningContent()
-	contentBlocks := agent.getLastAnthropicContentBlocks()
-	openAIToolCalls := buildOpenAIToolCallsForHistory(toolCalls)
-	appendAssistantToolCallsHistoryMessage(agent, explanation, reasoningContent, contentBlocks, openAIToolCalls)
-	if agent.Stats != nil {
-		agent.Stats.AssistantMessages++
-	}
-}
-
-func appendHeadlessToolResultToHistory(agent *Agent, toolCall *tools.ToolCall, result string) {
-	if agent == nil || toolCall == nil {
-		return
-	}
-	if !keepToolResultHistory(toolCall) {
-		return
-	}
-	agent.History = append(agent.History, toolruntime.BuildToolResultMessage(toolCall, result, toolruntime.FormatTextToolResultContent(toolCall.Tool, result)))
-}
-
-// extractToolFilePath はツール呼び出しから表示用ターゲットを抽出する。
-func extractToolFilePath(tc *tools.ToolCall) string {
-	if tc == nil {
-		return ""
-	}
-
-	switch tc.Tool {
-	case "gather_context":
-		if query := tc.Args["query"]; query != "" {
-			return query
-		}
-		if path := tc.Args["path"]; path != "" {
-			return path
-		}
-	case "read_file", "write_file", "str_replace", "delete_file", "list_dir", "lint", "format":
-		if path := tc.Args["path"]; path != "" {
-			return path
-		}
-	case "search_code":
-		if pattern := tc.Args["pattern"]; pattern != "" {
-			return fmt.Sprintf("%q", pattern)
-		}
-		if path := tc.Args["path"]; path != "" {
-			return path
-		}
-	case "bash":
-		if cmd := tc.Args["command"]; cmd != "" {
-			return truncateEventOutput(cmd, 40)
-		}
-	}
-
-	if rawPaths := tc.Args["paths"]; rawPaths != "" {
-		var paths []string
-		if err := json.Unmarshal([]byte(rawPaths), &paths); err == nil && len(paths) > 0 {
-			if len(paths) == 1 {
-				return paths[0]
-			}
-			return fmt.Sprintf("%s (+%d files)", paths[0], len(paths)-1)
-		}
-	}
-	if path := tc.Args["path"]; path != "" {
-		return path
-	}
-	if pattern := tc.Args["pattern"]; pattern != "" {
-		return pattern
-	}
-	if symbol := tc.Args["symbol"]; symbol != "" {
-		return symbol
-	}
-	return ""
-}
-
-// truncateEventOutput はイベント出力を制限する。
-func truncateEventOutput(s string, maxLen int) string {
-	if maxLen <= 0 {
-		return ""
-	}
-	runes := []rune(s)
-	if len(runes) <= maxLen {
-		return s
-	}
-	return string(runes[:maxLen]) + "..."
 }
